requirements.txt
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.7
cycler==0.12.1
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.128.8
fastapi-cli==0.0.24
fastapi-cloud-cli==0.17.0
fastar==0.11.0
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.3.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.4.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.9.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
narwhals==2.19.0
numpy==2.4.4
packaging==26.0
pandas==3.0.2
passlib==1.7.4
pillow==12.1.0
pluggy==1.6.0
protobuf==7.34.1
pyarrow==23.0.1
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.1
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
pygame==2.6.1
Pygments==2.19.2
PyMySQL==1.1.2
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.26
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.1
sentry-sdk==2.58.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
SQLAlchemy==2.0.49
starlette==0.52.1
streamlit==1.56.0
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
unicorn==2.1.4
urllib3==2.6.3
uvicorn==0.40.0
watchdog==6.0.0
watchfiles==1.1.1
websockets==16.0
